moving Choice page assets
