+++ Frontend/src/pages/Trash/Modify/index.jsx
@@ -1,18 +1,18 @@
 import "./modify.style.scss";
 // img
-import CertificateImg from "../../assets/img/cert.svg";
+import CertificateImg from "../../../assets/images/Choice/cert.svg";
 
 import Sidebar from "./Sidebar";
 
 const Modify = () => {
-	return (
-		<div className="modifyContainer">
-			<Sidebar />
-			<div className="certificate">
-				<img src={CertificateImg} alt="certicificate" />
-			</div>
-		</div>
-	);
+  return (
+    <div className="modifyContainer">
+      <Sidebar />
+      <div className="certificate">
+        <img src={CertificateImg} alt="certicificate" />
+      </div>
+    </div>
+  );
 };
 
 export default Modify;
